Change return type to size_t

+++ src/pd.c
@@ -275,15 +275,15 @@ int pd_write_file(const char *file_name, const pd_node_t **nodes) {
     return 0;
 }
 
-int pd_get_max_group_size() {
+size_t pd_get_max_group_size() {
     return PD_MAX_GROUP_SIZE;
 }
 
-int pd_get_max_key_size() {
+size_t pd_get_max_key_size() {
     return PD_MAX_KEY_SIZE;
 }
 
-int pd_get_max_value_size() {
+size_t pd_get_max_value_size() {
     return PD_MAX_VALUE_SIZE;
 }
